+++ src/board/board/board.js
@@ -1,8 +1,7 @@
 // more of a placeholder for now -
 // should talk when merged to properly integrate this with
 // existing "board" and "players" implementations if any exist
-function Board(players) {
-  this.players = players;
+function Board() {
   this.spaces = [];
 }
